refactor(app-router): React Router v6.4 방식으로 교체
Router 바깥에서 useLocation을 호출할 수 없어 일단 ScrollToTop을 무력화하였습니다.
어차피 ScrollToTop은 곧 ScrollRestoration을 위해 삭제될 예정이었습니다.

See also #319
