+++ app/src/App.tsx
@@ -1,15 +1,16 @@
 import { Global } from '@emotion/react';
 
 import { AppProvider } from '@core/AppProvider';
-import { AppRoutes } from '@core/AppRoutes';
+import { appRouter } from '@core/appRouter';
 import { global } from '@core/styles/global';
 import { reset } from '@core/styles/reset';
+import { RouterProvider } from 'react-router-dom';
 
 const App = () => {
   return (
     <AppProvider>
       <Global styles={[reset, global]} />
-      <AppRoutes />
+      <RouterProvider router={appRouter} />
     </AppProvider>
   );
 };
